Add, modified and changed namei
Added everything except the dry leaf weight for the lma of juniperus
species. Changed name of few folders and modified the spelling and
species id of some of the first trip data since I was not sure about one
of the species and I collected some new species in the second trip.
Overall fixed some typo, added data of the second trip and changed named
some files and folder