Added pre_burning_disc temp on model
The pre_burning_temperature added to the models
to see if whether it improved the models or not.
Created separate models to see if the remaining
less important traits have any significant effect on
repsonse variables or not. Fixed some minor issue e.g-
typo, created new variables in read_data.R