Updated scripts and README file
Couple of changes made in the script. Should have done
multiple commit. This is the least number of Rscript I can have for my
whole project. the ms_figures.R script is supposed to run now. I need to
some time to update rest of the task related to R.