Modified read_hobos.R
Modified the read_hobos.R in oder to read all the trials. Otherwise
I can't read the herbivore_analysis.R because I am using almost all the
samples from 2022 to do the herbivore part. Last time I did it to only
read the number of samples I need to do the analysis for 2022.