Fixed the pca plot
Fixed the pca plot in ms_figures.R. Previously "PC1" and " PC2" wer showed up as another variables in the plot. The reason was the the data set was used to create the plot contained the PC1 and PC2 which I did earlier. This plot is now after renaming the variables ,.