Add comment re hard-coding linear model values
This is fragile. Also, some figures use what seem to be z-score
transformed values (Eg fig1 x axis?).  Back-transform these so that the
reader sees the actual values.  If there was a log transformation before
modeling, do show data on a log scale, but z-scores are a purely
numerical stability transformation so back those out for presentation.